fix(Dmailer): Properly mark mail as sent after an error is resolved (refs #44815)
When sending an e-mail fails once and succeeds later, the successful
attempt was not properly marked (html_sent > 0). Therefore, a new
mail was generated for each remaining cycle for the respective
sys_dmail. This was fixed.

Apart from that, two maillog entries were generated in case of an error.
This was fixed, too.